When comboBox changed, loads the image corresponding to that layout. Add a red point to mark the begining of the stroke
Fill a circle of color red to mark of the begin of the stroke to make the word in the keyboard. Load one name for the .txt and .png of the keyboard layout to the comboBox. Added some lines to the README file, how the layouts needs to be read and the size and resolution of the keyboard image.